requirements.txt
azure-functions
pandas
scikit-learn
numpy
joblib
